+++ lib/src/oauth_token.dart
@@ -4,14 +4,15 @@ class OAuthToken {
   final String accessToken;
   final String? refreshToken;
   final DateTime? expiration;
+  final String? idToken;
 
-  bool get isExpired =>
-      expiration != null && DateTime.now().isAfter(expiration!);
+  bool get isExpired => expiration != null && DateTime.now().isAfter(expiration!);
 
   const OAuthToken._(
     this.accessToken,
     this.refreshToken,
     this.expiration,
+    this.idToken,
   );
 
   factory OAuthToken.fromJson(String json) {
@@ -23,5 +24,6 @@ class OAuthToken {
         credentials.accessToken,
         credentials.refreshToken,
         credentials.expiration,
+        credentials.idToken,
       );
 }